pyproject.toml
[tool.poetry]
name = "receipt-manager"
version = "0.1.0"
description = ""
authors = ["Yevhen Syzintsev <yevhen_s@dataforest.ai>"]
readme = "README.md"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
pydantic = "^2.10.6"
uvicorn = "^0.34.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.5"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.38"
bcrypt = "^4.2.1"
itsdangerous = "^2.2.0"
httpx = "^0.28.1"
asyncpg = "^0.30.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-order = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
